Patch from Chico Charlesworth addressing issue 1621, "Define Lookup Options" does not support adding new Collateral Type entries (https://mifos.dev.java.net/issues/show_bug.cgi?id=1621)
Includes changes for mapping default Collateral Types with ids 1 & 2 to corresponding LOOKUP_VALUE entries with ids 109 and 110.

An additional change was made to CustomValueListElement to handle core i18n changes so that retreiving the "name" property will work correctly.




git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@12553 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/src/org/mifos/application/accounts/loan/struts/actionforms/LoanAccountActionForm.java
@@ -442,8 +442,8 @@ public class LoanAccountActionForm extends BaseActionForm {
 		return getIntegerValue(getBusinessActivityId());
 	}
 
-	public Short getCollateralTypeIdValue() {
-		return getShortValue(getCollateralTypeId());
+	public Integer getCollateralTypeIdValue() {
+		return getIntegerValue(getCollateralTypeId());
 	}
 
 	public Money getLoanAmountValue() {
